Implements Auth0 React tutorial
Followed Auth0 React tutorial exactly to get Auth0 configured within app. Like the tutorial, we only used the 3 basic views to get started. There were issues with using withRouter, so we decided to keep the tutorial's history component for redirects for now. At this point, the Auth0 changes need to be integrated with the rest of the original app's functionality.

Pair-programming partner: Drew Moody

 Changes to be committed:
	modified:   .env
	new file:   callback-cross-auth.html
	modified:   package.json
	modified:   public/index.html
	modified:   src/App.js
	new file:   src/auth/Auth.js
	new file:   src/auth/Callback.js