Front end makes axios calls to backend with auth token
Changed audience argument in checkJwt to aud as per a GitHub issuerecommendation, which allows the front-end to communicate with the back-end using token validation. Implements a simple axios call in place of the dummy data call in App.js. The authorization header requires a bearer token for the back-end to validate using checkJwt. The audience property in the auth config must be set correctly.

 Changes to be committed:
	modified:   ../spaced-repetition-back-end/jwt.js
	modified:   ../spaced-repetition-back-end/routes/users/usersRoutes.js
	modified:   src/App.js
	modified:   src/auth/Auth.js