User profile functionality in global store and component
On redirction to the dashboard, the authenticated user's profile is fetched and set in the global state in App.js. When navigating to `/dashboard/profile` the component fetches the user's profile if it is not already cached in the auth instance. The Profile component is still a WIP. The basic setup is there to build from.

 Changes to be committed:
	modified:   src/App.js
	modified:   src/auth/Auth.js
	modified:   src/components/Profile.js
	modified:   src/components/Wrapper.js