Landing page redesign and reset now working
Pushing up because reset was not working before. For the sake of time it is in a .css file now until we get it working with styled components. It is not a priority to be in styled components. The landing page structure is there for the header, pricing, and team. Still needs responsiveness. The reset also changed some of the styling. Other teammates need the new code because of the reset.